anc-with-ai: fix(dtype): normalize int16 capture to float32 not float64 [phase5]
Mirror of the anc-backend fix (near-identical fork): audio_capture.py and
predict_noise_type.py cast int16 PCM to float64, escaping the I/O boundary and
violating the float32 [-1,1] contract. Surfaced by the Phase 5 dtype audit.

+++ src/ml/predict_noise_type.py
@@ -79,7 +79,7 @@ class NoisePredictor:
         with wave.open(filepath, 'rb') as wf:
             frames = wf.readframes(wf.getnframes())
             audio_data = np.frombuffer(frames, dtype=np.int16)
-            audio_data = audio_data.astype(np.float64) / 32768.0
+            audio_data = audio_data.astype(np.float32) / 32768.0
 
         return self.predict_from_audio(audio_data)
